pipeline-training/requirements.txt
functions-framework==3.1.0
google-cloud-aiplatform==1.16.1
PyYAML==6.0
